+++ server/DocumentSyncHandler.cs
@@ -1,3 +1,5 @@
+namespace TScriptLanguageServer;
+
 using MediatR;
 using OmniSharp.Extensions.LanguageServer.Protocol;
 using OmniSharp.Extensions.LanguageServer.Protocol.Models;
@@ -7,8 +9,6 @@ using OmniSharp.Extensions.LanguageServer.Protocol.Document;
 using OmniSharp.Extensions.LanguageServer.Protocol.Client.Capabilities;
 using OmniSharp.Extensions.LanguageServer.Protocol.Server.Capabilities;
 using OmniSharpRange = OmniSharp.Extensions.LanguageServer.Protocol.Models.Range;
-
-namespace TScriptLanguageServer;
 
 public class DocumentSyncHandler : ITextDocumentSyncHandler
 {