+++ server/Main.cs
@@ -1,12 +1,31 @@
+namespace TScriptLanguageServer;
+
 using System.Reflection;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.DependencyInjection;
 using OmniSharp.Extensions.LanguageServer.Server;
-
-namespace TScriptLanguageServer;
+using TScriptLanguageServer.Language;
 
 public class Program
 {
+    // public static void DebuggingMain(string[] args)
+    // {
+    //     string? currentDir = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
+    //     string logFileName = "tscript.language.server.log";
+    //     string logFilePath = Path.Combine(currentDir ?? "", logFileName);
+    //     Logging.OutputStream = new(logFilePath);
+    //     Logging.DebugMessagesEnabled = true;
+
+    //     Tokenizer tokenizer = new();
+    //     string buffer = File.ReadAllText("audio.tscript");
+    //     tokenizer.ParseFile(buffer);
+
+    //     TokenParser p = new();
+    //     p.Parse(tokenizer);
+
+    //     Logging.OutputStream.Close();
+    // }
+
     public static async Task Main(string[] args)
     {
 
@@ -14,11 +33,14 @@ public class Program
         string logFileName = "tscript.language.server.log";
         string logFilePath = Path.Combine(currentDir ?? "", logFileName);
         Logging.DebugMessagesEnabled = true;
-        if (args.Length == 2)
+        if (args.Length >= 1)
+            Specifications.StdLibPath = args[0];
+
+        if (args.Length == 3)
         {
-            _ = bool.TryParse(args[0], out Logging.DebugMessagesEnabled);
-            if (!string.IsNullOrEmpty(args[1]))
-                logFilePath = args[1];
+            _ = bool.TryParse(args[1], out Logging.DebugMessagesEnabled);
+            if (!string.IsNullOrEmpty(args[2]))
+                logFilePath = args[2];
         }
         Logging.OutputStream = new(logFilePath);
         for (int i = 0; i < args.Length; i++)
@@ -34,6 +56,7 @@ public class Program
                     .SetMinimumLevel(LogLevel.Debug))
                 .WithHandler<DocumentSyncHandler>()
                 .WithHandler<ScriptCompletionHandler>()
+                .WithHandler<TooltipHandler>()
                 .WithServices(ConfigureServices)
             );
             await server.WaitForExit;